+++ changelog.md
@@ -1,3 +1,6 @@
+### November 19th, 2020
+#### @ayx/ayx-ui-sdk@0.0.5-alpha
+- Added additional support for multiple encryption types
 ### November 9th, 2020
 #### @ayx/ayx-ui-sdk@0.0.4-alpha
 - Changes to sonarqube integration to correctly report numbers